chore: fix flaky test - attempt #2
I tried increasing retries here: https://github.com/global-121/121-platform/pull/5240

But it seems that is not enough because it still fails 3 times in a row at times.

Now I'm trying to increase the timeouts, and increase frequency of checking, in the relevant helpers, with the hope that this is caused by a slow inner network in CI (at times).