chore: improve changePreferredLanguage test
In my branch, this test was failing, but I couldn't tell why.

Then I realised that the test was proceeding, even though an error occurred. That is because it pressed "ok" twice even if there was an error, without checking the label of what it was clicking "ok" on.

This should improve the messaging when it actually fails.