chore: create template for automatically generated release notes
I was wondering whether it was possibly to configure this in some way. I found this: https://docs.github.com/en/repositories/releasing-projects-on-github/automatically-generated-release-notes#example-configurations

I figured I would give it a try while we figure out what we do with the actual changelog.

My 0.02$: I would remove CHANGELOG.md altogether and just use the release notes on the releases themselves, and allow github to generate them for us.