chore: remove CHANGELOG
I don't see the value of maintaining the CHANGELOG, especially when we can achieve the same by posting the updates directly in the release notes of the https://github.com/global-121/121-platform/releases

Having to maintain the changelog adds this extra step during the release process which requires an additional approval and IMO doesn't add any value.

The file I've added at `.github/release.yml` will automatically create a release note based on the PRs merged in the release. We can (optionally) use the labels on our PRs to organise the release notes.