portalicious: better error for change-password (#5716)
portalicious:  better error for change-password

AB#29793

This does a few things:
- Moves the form/backend error to be closer to the button rather than closer to the fields
- Removes the custom error layer from auth.service
- Changes the default error handling from http-wrapper. Usually the BE sends a message together with a status code. If that is present, we use that instead of overriding with a custom message.

I'm not sure if this is the best fix. That last point means that a part of the error will never be localised. I think this is a good tradeoff, because at least the error is "always" more precise, but I'm not sure.