portalicious: edit user on teams page
AB#29744

When discussing with Elwin, we had settled on a different approach:
The idea was to get rid of the separate component for managing the form and moving everything into the page.

When I started doing this, though, I realised that maybe
there wasn't a need for it, and instead the only logic that
needed to be moved to the page was whether to show the form or not.

So, I've also edited the "create-project-button" component accordingly so that they both follow the same pattern.

Still not sure if this is the wisest approach but it works.