portalicious: add registrations breadcrumb and handle not found
AB#30165
AB#30395

I combined these two tasks because I found one relying on the other, and figured that they wouldn't be a large set of changes even when combined.

The design of the "registration not found" deviates from what is in Figma based on a convo with Tal.

I also took the liberty of adding a "project not found" error.