portalicious: remove Dto121Service
AB#31183

The issues we are having on feature branches are related to the use of Dto121Service, where we import directly from the controllers, and end up (for reasons unbeknownst to me) in a situation where TS can't resolve these kinds of imports:

```bash
~/git/121-platform/interfaces/Portalicious ➜ npm run typecheck

> typecheck
> npx tsc --noEmit

../../services/121-service/src/registration/repositories/registration-scoped.repository.ts:14:45 - error TS2307: Cannot find module 'typeorm/query-builder/QueryPartialEntity' or its corresponding type declarations.

14 import { type QueryDeepPartialEntity } from 'typeorm/query-builder/QueryPartialEntity';
                                               ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

../../services/121-service/src/scoped.repository.ts:16:40 - error TS2307: Cannot find module 'typeorm/query-builder/QueryPartialEntity' or its corresponding type declarations.

16 import { QueryDeepPartialEntity } from 'typeorm/query-builder/QueryPartialEntity';
                                          ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
Found 2 errors in 2 files.
```

I've decided to remove Dto121Service for now, and consistently rely only on dtos, enums, and entities from the 121-service. These seem to have less issues becuase the dependencies in those files don't trigger these weird warnings.

We can figure out how to introduce it when we pick up AB#30152