Use `loadComponent` in app.routes.ts
This change addresses an issue where the registrations table
component was not working properly when being used in two
different places in the application. Angular was spitting
out a very cryptic error on the registrations page:

```
Cannot read properties of undefined (reading 'ɵcmp')
```

When looking for solutions, this was mentioned as a fix.

I tried it and it works, fixing the E2E tests. I also
noticed that this was already the strategy being used
in the old portal so I guess it's a good idea to keep it.