portalicious: consistent "esc" behaviour on dialogs
AB#31449

Given that sometimes we don't want `esc` to close dialogs, this makes it consistent for all.

I couldn't find a way to set this "globally" for all components - without wrapping the primeng component but IMHO that is overkill for a single attribute.

It would've been nice to enforce this via eslint, but it depnds on this proposal:
https://github.com/angular-eslint/angular-eslint/discussions/1800

I tried doing it with the html eslint plugin, but I don't think
that can co-exist with the angular-eslint plugin.

For now we'll have to track this manually I suppose.