portalicious: fix change status "cancel" button (#6448)
AB#33160

I checked out the original commit where these modals were added:
49d028e4a4e38c008c2d9c6ca82fd48939fca6ff

And it seems this was always broken. I remember that, when we built this originally, me and Peter went through several refactors. I suppose this slipped through the cracks at the time.

The fix is mainly the `(cancelClick)="cancelClick()"` change. The other change is a design fix related to the same buttons.