portalicious: fix validation on change password page (#6449)
AB#30007

I had a look at a lot of forms, and I think with the theming efforts the consistency around validation is there now.

This was the only scenario that was a bit funky, but it's also the only page where we're using ValidatorFn (because of multi-field validation) so it's needs this extra nudge.