chore: add knip to portalicious
AB#34711

I wanted to add knip to avoid dead code such as this:
https://github.com/global-121/121-platform/pull/6583/files#r2007873570

Turns out, there isn't a tool capable of detecting unused public class properties in tyepscript:
https://github.com/microsoft/TypeScript/issues/29293

So it doesn't solve the use case I had for it. But still, we have it enabled on the 121-service, so I figured why not enable it here too and get rid of some dead code.