chore: check for dead code in portalicious with knip (#6638)
* chore: add knip to portalicious

AB#34711

I wanted to add knip to avoid dead code such as this:
https://github.com/global-121/121-platform/pull/6583/files#r2007873570

Turns out, there isn't a tool capable of detecting unused public class properties in tyepscript:
https://github.com/microsoft/TypeScript/issues/29293

So it doesn't solve the use case I had for it. But still, we have it enabled on the 121-service, so I figured why not enable it here too and get rid of some dead code.

* chore: Remove unused code for handling server-side translations via Transifex (#6642)

* Remove transifex js

Signed-off-by: Ruben <vandervalk@geoit.nl>

* chore: Remove unused dependencies of unused code (#6643)

* fix: Remove unused yargs dependencies

* fix: Remove unused tsc-watch, as we only use the Nest.js CLI to watch

---------

Signed-off-by: Ruben <vandervalk@geoit.nl>
Co-authored-by: Ruben <vandervalk@geoit.nl>
Co-authored-by: Elwin Schmitz <eschmitz@redcross.nl>

* chore: add knip to portalicious

AB#34711

I wanted to add knip to avoid dead code such as this:
https://github.com/global-121/121-platform/pull/6583/files#r2007873570

Turns out, there isn't a tool capable of detecting unused public class properties in tyepscript:
https://github.com/microsoft/TypeScript/issues/29293

So it doesn't solve the use case I had for it. But still, we have it enabled on the 121-service, so I figured why not enable it here too and get rid of some dead code.

---------

Signed-off-by: Ruben <vandervalk@geoit.nl>
Co-authored-by: Elwin Schmitz <eschmitz@redcross.nl>
Co-authored-by: RubenGeo <34537157+RubenGeo@users.noreply.github.com>
Co-authored-by: Ruben <vandervalk@geoit.nl>