chore: fix flaky ExportPayments test
AB#38159

In #7106 the payments export changed the name of the payment ID column from `payment` to `paymentId`.

The snapshot was updated to reflect this change, but a piece of the test code was still relying on sorting the records based on `payment`.

This sort function was silently failing, introducing flakiness on the test.

Besides fixing the sort function, I've also added assertions to make sure that a change in column name would not go unnoticed again.