chore: better error handling in portal
AB#38599

Spin-off from #7383

Whilst working on #7383, I realised some errors from the BE
were being swallowed in the frontend. An easy fix was to
introduce this `JSON.stringify` in the error handler.

When the errorMessage we've extracted is a string, it just adds
double quotes around it, but then when it is an object it will
output the contents of the object, rather than
`[Object object]` which is what I was seeing before.